refactor error handling
