--- ocean/ocean-api-core/src/errors/ApiError.ts
@@ -1,19 +1,0 @@
-export enum ApiErrorType {
-  ValidationError = 'ValidationError',
-  BadRequest = 'BadRequest',
-  NotFound = 'NotFound',
-  Conflict = 'Conflict',
-  Forbidden = 'Forbidden',
-  Unauthorized = 'Unauthorized',
-  BadGateway = 'BadGateway',
-  TimeoutError = 'TimeoutError',
-  UnknownError = 'UnknownError',
-}
-
-export interface ApiError {
-  code: number
-  type: ApiErrorType
-  at: number
-  message?: string
-  url?: string
-}
