+++ ocean/ocean-api-server/src/RootModule.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { RootConfiguration } from './RootConfiguration'
 import { ControllerModule } from './controllers/_Module'
-import { ServiceModule } from './services/_Module'
 
 @Module({
   imports: [
@@ -11,7 +10,6 @@ import { ServiceModule } from './services/_Module'
       load: [RootConfiguration]
     }),
     ControllerModule,
-    ServiceModule,
   ]
 })
 export class RootModule {
