+++ ocean/ocean-api-server/src/controllers/_Module.ts
@@ -2,7 +2,7 @@ import { CacheModule, Module } from '@nestjs/common'
 import { ActuatorController } from './ActuatorController'
 import { FeeController } from './FeeController'
 import { APP_INTERCEPTOR, APP_PIPE } from '@nestjs/core'
-import { ApiValidationPipe } from './filters/ApiValidationPipe'
+import { GlobalValidationPipe } from './filters/GlobalValidationPipe'
 import { ResponseInterceptor } from './filters/ResponseInterceptor'
 import { ExceptionInterceptor } from './filters/ExceptionInterceptor'
 import { ConfigService } from '@nestjs/config'
@@ -20,9 +20,8 @@ import { NetworkName } from '@defichain/jellyfish-network'
   providers: [
     {
       provide: APP_PIPE,
-      useClass: ApiValidationPipe
+      useClass: GlobalValidationPipe
     },
-    // APP_INTERCEPTOR are only activated for /v* paths
     {
       provide: APP_INTERCEPTOR,
       useClass: ResponseInterceptor
