+++ ocean/ocean-api-server/src/controllers/filters/ResponseInterceptor.ts
@@ -11,10 +11,6 @@ import { ApiPagedResponse, ApiResponse } from '@defichain/ocean-api-core/src'
 @Injectable()
 export class ResponseInterceptor<T> implements NestInterceptor<T> {
   intercept (context: ExecutionContext, next: CallHandler<T>): Observable<any> {
-    if (!isVersionPrefixed(context)) {
-      return next.handle()
-    }
-
     return next.handle().pipe(map((result): ApiResponse<any> => {
       if (result instanceof ApiPagedResponse) {
         return {
@@ -26,14 +22,4 @@ export class ResponseInterceptor<T> implements NestInterceptor<T> {
       return { data: result }
     }))
   }
-}
-
-/**
- * @param {ExecutionContext} context to check if path is version prefixed
- * @return {boolean}
- */
-export function isVersionPrefixed (context: ExecutionContext): boolean {
-  const request = context.switchToHttp().getRequest()
-  const url: string = request.raw?.url
-  return url.startsWith('/v')
 }
