--- ocean/ocean-api-server/src/services/_Module.ts
@@ -1,5 +1,0 @@
-import { Module } from '@nestjs/common'
-
-@Module({})
-export class ServiceModule {
-}
